+++ bindings/python/examples/how_tos/accounts_and_addresses/consolidate_outputs.py
@@ -34,13 +34,9 @@ print('Outputs BEFORE consolidation:')
 
 for i, output_data in enumerate(outputs):
     print(f'OUTPUT #{i}')
-    print(
-        '- address: {}\n- amount: {}\n- native tokens: {}'.format(
-            Utils.hex_to_bech32(output_data.address.pubKeyHash, 'rms'),
-            output_data.output.amount,
-            output_data.output.nativeTokens
-        )
-    )
+    print(f'- address: #{Utils.hex_to_bech32(output_data.address.pubKeyHash, "rms")}')
+    print(f'- amount: #{output_data.output.amount}')
+    print(f'- native tokens: #{output_data.output.nativeTokens}')
 
 print('Sending consolidation transaction...')
 
@@ -66,10 +62,6 @@ outputs = account.unspent_outputs()
 print('Outputs AFTER consolidation:')
 for i, output_data in enumerate(outputs):
     print(f'OUTPUT #{i}')
-    print(
-        '- address: {}\n- amount: {}\n- native tokens: {}'.format(
-            Utils.hex_to_bech32(output_data.address.pubKeyHash, 'rms'),
-            output_data.output.amount,
-            output_data.output.nativeTokens
-        )
-    )
+    print(f'- address: #{Utils.hex_to_bech32(output_data.address.pubKeyHash, "rms")}')
+    print(f'- amount: #{output_data.output.amount}')
+    print(f'- native tokens: #{output_data.output.nativeTokens}')