+++ bindings/python/iota_sdk/__init__.py
@@ -1,13 +1,13 @@
 # Copyright 2023 IOTA Stiftung
 # SPDX-License-Identifier: Apache-2.0
 
-from .iota_sdk import *
 from .client.client import Client, NodeIndexerAPI, ClientError
 from .client._high_level_api import GenerateAddressesOptions, GenerateAddressOptions
+from .external import *
 from .utils import Utils
 from .wallet.wallet import Wallet, Account
 from .wallet.common import WalletError
-from .wallet.sync_options import AccountSyncOptions, AliasSyncOptions, NftSyncOptions, SyncOptions
+from .wallet.sync_options import AccountSyncOptions, NftSyncOptions, AliasSyncOptions, SyncOptions
 from .secret_manager.secret_manager import *
 from .prefix_hex import *
 from .types.address import *