+++ bindings/python/iota_sdk/wallet/prepared_transaction.py
@@ -2,10 +2,10 @@
 # SPDX-License-Identifier: Apache-2.0
 
 from __future__ import annotations
+from typing import TYPE_CHECKING, Dict, Union
+from dacite import from_dict
 from iota_sdk.types.transaction import Transaction
 from iota_sdk.types.transaction_data import PreparedTransactionData
-from dacite import from_dict
-from typing import TYPE_CHECKING, Dict, Union
 # Required to prevent circular import
 if TYPE_CHECKING:
     from iota_sdk.wallet.wallet import Account
@@ -29,24 +29,11 @@ class PreparedTransaction:
         self.account = account
         self.prepared_transaction_data_dto = prepared_transaction_data
 
-    """
-    The function returns the prepared transaction data.
-
-    :returns: The method prepared_transaction_data() is returning an object of type PreparedTransaction
-    """
-
     def prepared_transaction_data(self) -> PreparedTransactionData:
-        """Get the prepared transaction data.
+        """Returns the prepared transaction data.
         """
         return self.prepared_transaction_data_dto if isinstance(
             self.prepared_transaction_data_dto, PreparedTransactionData) else from_dict(PreparedTransactionData, self.prepared_transaction_data_dto)
-
-    """
-    The send function returns a promise that resolves to a Transaction object after signing
-    and submitting the transaction. Internally just calls sign_and_submit_transaction.
-
-    :returns: The send() method is returning a Transaction object after it has been signed and submitted.
-    """
 
     def send(self) -> Transaction:
         """Send a transaction. Internally just calls `sign_and_submit_transaction`.
@@ -76,7 +63,6 @@ class PreparedTransaction:
 class PreparedCreateTokenTransaction(PreparedTransaction):
 
     """A prepared transaction for creating a native token.
-    The function returns the token_id as a string.
 
     Returns: The token id of the PreparedCreateTokenTransaction.
     """
@@ -87,7 +73,7 @@ class PreparedCreateTokenTransaction(PreparedTransaction):
         return self.prepared_transaction_data_dto["tokenId"]
 
     def prepared_transaction_data(self):
-        """The function returns the prepared transaction data.
+        """Returns the prepared transaction data.
         """
         return from_dict(PreparedTransactionData,
                          self.prepared_transaction_data_dto["transaction"])