+++ bindings/python/iota_sdk/wallet/wallet.py
@@ -1,14 +1,15 @@
 # Copyright 2023 IOTA Stiftung
 # SPDX-License-Identifier: Apache-2.0
 
+from json import dumps
+from typing import Any, Dict, List, Optional, Union
 from iota_sdk import destroy_wallet, create_wallet, listen_wallet, get_client_from_wallet, get_secret_manager_from_wallet, Client
 from iota_sdk.secret_manager.secret_manager import LedgerNanoSecretManager, MnemonicSecretManager, StrongholdSecretManager, SeedSecretManager, SecretManager
-from iota_sdk.types.client_options import ClientOptions
 from iota_sdk.types.address import AccountAddress
 from iota_sdk.wallet.account import Account, _call_method_routine
 from iota_sdk.wallet.sync_options import SyncOptions
-from json import dumps
-from typing import Any, Dict, List, Optional, Union
+
+# pylint: disable=too-many-public-methods
 
 
 class Wallet():
@@ -18,7 +19,10 @@ class Wallet():
         handle: The wallet handle.
     """
 
-    def __init__(self, storage_path: Optional[str] = None, client_options: Optional[Dict[str, Any]] = None, coin_type: Optional[int] = None,
+    def __init__(self,
+                 storage_path: Optional[str] = None,
+                 client_options: Optional[Dict[str, Any]] = None,
+                 coin_type: Optional[int] = None,
                  secret_manager: Optional[Union[LedgerNanoSecretManager, MnemonicSecretManager, SeedSecretManager, StrongholdSecretManager]] = None):
         """Initialize `self`.
         """
@@ -273,9 +277,9 @@ class Wallet():
         """
         return destroy_wallet(self.handle)
 
+    # pylint: disable=redefined-builtin
     @staticmethod
     def __return_str_or_none(str):
         if str:
             return str
-        else:
-            return None
+        return None