+++ bindings/wasm/src/wallet.rs
@@ -50,9 +50,10 @@ pub async fn destroy_wallet(method_handler: &WalletMethodHandler) -> Result<(), 
 
 #[wasm_bindgen(js_name = getClientFromWallet)]
 pub async fn get_client(method_handler: &WalletMethodHandler) -> Result<ClientMethodHandler, JsValue> {
-    let wallet = method_handler.wallet.lock().await;
-
-    let client = wallet
+    let client = method_handler
+        .wallet
+        .lock()
+        .await
         .as_ref()
         .ok_or_else(|| "wallet got destroyed".to_string())?
         .client()
@@ -63,9 +64,10 @@ pub async fn get_client(method_handler: &WalletMethodHandler) -> Result<ClientMe
 
 #[wasm_bindgen(js_name = getSecretManagerFromWallet)]
 pub async fn get_secret_manager(method_handler: &WalletMethodHandler) -> Result<SecretManagerMethodHandler, JsValue> {
-    let wallet = method_handler.wallet.lock().await;
-
-    let secret_manager = wallet
+    let secret_manager = method_handler
+        .wallet
+        .lock()
+        .await
         .as_ref()
         .ok_or_else(|| "wallet got destroyed".to_string())?
         .get_secret_manager()
@@ -79,10 +81,19 @@ pub async fn get_secret_manager(method_handler: &WalletMethodHandler) -> Result<
 /// Returns an error if the response itself is an error or panic.
 #[wasm_bindgen(js_name = callWalletMethodAsync)]
 pub async fn call_wallet_method_async(method: String, method_handler: &WalletMethodHandler) -> Result<String, JsValue> {
-    let wallet = method_handler.wallet.lock().await;
     let method: WalletMethod = serde_json::from_str(&method).map_err(|err| err.to_string())?;
 
-    let response = call_wallet_method(wallet.as_ref().expect("wallet got destroyed"), method).await;
+    let response = call_wallet_method(
+        method_handler
+            .wallet
+            .lock()
+            .await
+            .as_ref()
+            .expect("wallet got destroyed"),
+        method,
+    )
+    .await;
+
     match response {
         Response::Error(e) => Err(e.to_string().into()),
         Response::Panic(p) => Err(p.into()),