+++ sdk/src/client/secret/private_key.rs
@@ -47,9 +47,7 @@ impl SecretManage for PrivateKeySecretManager {
         let public_key = self.0.public_key().to_bytes();
 
         // Hash the public key to get the address
-        let result = Blake2b256::digest(public_key).try_into().map_err(|_e| {
-            crate::client::Error::Blake2b256("hashing the public key while generating the address failed.")
-        })?;
+        let result = Blake2b256::digest(public_key).into();
 
         crate::client::Result::Ok(vec![Ed25519Address::new(result)])
     }