+++ sdk/tests/types/unlock/nft.rs
@@ -3,6 +3,7 @@
 
 use iota_sdk::types::block::{unlock::NftUnlock, Error};
 use packable::{bounded::InvalidBoundedU16, PackableExt};
+use pretty_assertions::assert_eq;
 
 #[test]
 fn kind() {