+++ sdk/tests/wallet/native_tokens.rs
@@ -5,6 +5,7 @@ use iota_sdk::{
     wallet::{account::SyncOptions, CreateNativeTokenParams, Result},
     U256,
 };
+use pretty_assertions::assert_eq;
 
 use crate::wallet::common::{create_accounts_with_funds, make_wallet, setup, tear_down};